Remove the support of patching "std::optional"
Summary:
This feature itself is inconsistent -- it specifically only works with `std::optional`, but not other optional types such as `folly::Optional`. It also seems a bit over-engineering since `std::optional` can never be a thrift field.

In this diff I changed such usage to a build failure. If this doesn't cause any actual build failure, we can assume this is never used in production and we can safely remove it.

Reviewed By: iahs

Differential Revision: D46215245

fbshipit-source-id: 3f96bde138454f8f3da0b5bae798df603b7f2db4