Remove deterministic_hash(obj)
Summary:
`deterministic_hash(obj)` uses StdHasher by default, which is bit misleanding.

It only guarantees to "produce the same result for the same input within **a single execution of a program**" (https://en.cppreference.com/w/cpp/utility/hash). However, it gives the impression that it always produces the same result for the same input.

This diff removes it.

In the future, we can reintroduce `deterministic_hash` but use an actual deterministic algorithm such as xxhash.

Reviewed By: thedavekwon

Differential Revision: D46250782

fbshipit-source-id: d24418a7fae52dfeca19257ba5344afaea644c54