bugfix: ensure after erase
Summary:
If we erased a key first, then ensure it later, the behavior of patch should be replacing the value of the key regardless whether key exists in map or not.

Thus, we need to not only ensure the key, but also replace the key in `patchAfter`.

The existing code doesn't take erase into account, thus the key will still be removed if we did `add(...)` after `erase(...)`.

Reviewed By: vitaut

Differential Revision: D46448997

fbshipit-source-id: e46d536461e259e09a32043601b1e08d5c9631ca