Add isAllMapMask() and isNoneMapMask() helper functions
Summary:
The current isAllMask()/isNoneMask() don't work for map masks. They give an illusion to users that it should work for all mask types, but they only work for field masks. They are misused in some places already.

To avoid breaking existing things, two new methods are added for map mask
types.

A better change would be to have following functhos
* isAllFieldMask()/isNoneFieldMask(): work for field mask, throw for map masks
* isAllMapMask()/isNoneMapMask(): work for map masks, throw for field mask
* isAllMask()/isNoneMask(): work for any mask type, it doesn't throw.

But this may need change in all places that depend on this.

Reviewed By: Mizuchi

Differential Revision: D45808732

fbshipit-source-id: 3a2fd1cae85a402b00665a5a1ea1c527192cf854