iobuf support adapted types
Summary:
This diff implements serialize and deserialize for adapted types. When a type-adapted type is `Extract`ed or `Construct`ed, the non-thrift  adapted type is converted to the thrift type with the adapter then serialized, or vv. Here "non-thrift" denotes that the class doesn't define a `read` or `write` method for serialization/deserialization.

There's also a tricky corner case involving a `Transitive` structured annotation definition used to propagate a field adapter type. Since this struct is a thrift type, it needs to be serialized/deserialized like normal. However, from the compiler's perspective, there's an adapter attached and no indication in the AST that it's a field adapter, not a type adapter. So we need SFINAE to handle the structured annotation definition.

Reviewed By: yfeldblum

Differential Revision: D46579843

fbshipit-source-id: fdc7c4ebb7746818e9fb8bf601dd5bd16fd9ab78