Test send unbatching & fragment handling for FDs
Summary:
This is a very direct analog of the same tests for server sends (D45507928). Copy-paste from that description:

The primary goal is to exercise the FD-based unbatching code -- that is, in the presence of FDs, we must break up the already-batched write into several socket writes. Here's what it checks:
- Separate messages carrying FDs are separated into separate socket invocations.
- Messages without FDs continue being batched together, and are batched with the subsequent FD-carrying message (if there is one in the batch).
- The socket is passed FDs together with the last fragment of a message.
- All the data & FDs enqueued into a batch are sent.
- The right data matches to the right FD in the socket code.

Reviewed By: AkramaMirza

Differential Revision: D45937794

fbshipit-source-id: ce9ab0859f0cd380e382b5efa2d3621d735d1c0f