Don't batch when FDs are being sent
Summary:
This is an exact analog of D43935129, but for client-sent FDs instead of server-sent. We break up write batches, so that each socket write has one message's FDs -- those of the last message in the batch.

See the tests on D45937794 or the design docs (`docs/contributions/fd-passing`) for the detailed invariants.

Reviewed By: AkramaMirza

Differential Revision: D43935126

fbshipit-source-id: 04e15d7cc6ff9d90d4c77e03b0a3c70ef0259ded