"marshal" empty structs; add opt-in annotation
Summary:
Added a struct-scope structured annotation to denote opt-in or opt-out. In the end state, this annotation is removed, but it allows us to test this on particular thrift structs for now without rolling out generally. Since "marshaling" an empty struct is trivial, this diff enables skipping call to serializer and just returning the empty struct, even without the opt-in annotation.

Added a type check as initial step of marshaling. This is probably the most graceful way to handle struct mismatch; without this, need to add manual field size checks before PyTuple_GET_ITEM calls in next diff: primitive marshaling.

Reviewed By: Filip-F

Differential Revision: D46411141

fbshipit-source-id: 9a0343bc326c894a97a61c03adf141ba7ba58b33