extract composed struct/enums
Summary:
Logic to marshal structs composed of other structs / enums. Also sync the fixture for bonus points.

Overall, this was *more* work than expected because thrift-python stores (a) the python integer for enums, and (b) the _fbthrift_data PyTuple for structs. That forced a bifurcation of `Extractor` for internal vs external. Luckily for enums it can be templated. For structs, the "inner" `Extractor` is separated out of the "external" `Extractor` with the latter calling the former for code re-use.

This diff only supports composition of in-module types. The logic to enable cross-module type composition is enabled in next diff to de-risk this diff.

Reviewed By: Filip-F

Differential Revision: D46715745

fbshipit-source-id: 937a82fc90eee4e27a1078336f70dfaf504a4acd