extract sets
Summary:
Implement sets marshaling.

The only interesting part of this diff is using PyIter protocol to loop through the set elements. `PyObject_Iter` and `PyIter_Next` return *new* references that must be DECREF'd to avoid leak.

Introduce `StrongRef` struct to do this RAII style with auto DECREF on drop. This is designed to be a temporary holder for "new" refs returned from CAPI functions, so it "steals" on construction and has a `release` method to pass ownership without DECREF. This will get more use when implementing Constructor (cpp -> py).

Reviewed By: Filip-F

Differential Revision: D46752698

fbshipit-source-id: 63fa9a975d468a6ecce52e7f31273f9244bf386e