Allow creation of zero-dependency thrift clients that skip event handler invocations
Summary:
Currently, all client-side `TProcessorEventHandler`s are stored in a global registry (`TClientBase::addEventHandler` and `TClientBase::addClientEventHandlerFactory`). This approach is necessarily for client objects because there is no "runtime" object to hold state, like `ThriftServer` for server-side event handlers.

This allows `AsyncClient` objects to be constructed with the option to bypass all globally registered event handlers. This means that all instances via `addEventHandler` will not be added, and `getEventHandler` won't be called on factory objects passed via `addClientEventHandlerFactory`.

Differential Revision: D46875424

fbshipit-source-id: f4a2666de54957ddce7d03e7d8a4f780e3385e9d