fix Extractor memory leak
Summary:
Added test to detect memory leak in `Extractor`. This works because primitive types are placed directly into struct internal data without conversion, so the same object can be round-tripped through the struct. In other words, the object put into struct field at construction `is` the same object returned by accessor for `int`, `float`, and `bytes`.

The leak happens because `getThriftData` uses `._fbthrift_data` in cython, which `INCREF`s. Therefore, need to put the return in a `StrongRef` to `DECREF` on scope exit.

Reviewed By: Filip-F

Differential Revision: D46957624

fbshipit-source-id: 392225e07a739a58e6cea2784f68829cbcb92f8c