switch thrift Ctors to const&
Summary: Switch enums and arithmetic types to val, everything else to `const&`. This saves copies in sets and maps, as well as for `Shared` -type `cpp.Ref` fields. It also saves a copy if the entry-point struct lives in `std::shared_ptr`, which is useful for converting `thrift-py3` struct to `thrift-python` using `Constructor`.

Reviewed By: yfeldblum

Differential Revision: D47044880

fbshipit-source-id: 1988efccdeabb03f8c5f13a4c72242d7f1209419