Stop passing around empty "persistentwriteheaders"
Summary:
These are misleading in that they look like they should contain persistentwriteheaders, which are an actual thing that can be set, but this object is quite literally never non-empty. We don't pull the info from anywhere, instead we just construct empty objects to satisfy function signatures.

Let's kill it.

Reviewed By: prasad223

Differential Revision: D46942165

fbshipit-source-id: 3faa59073b6827ab5e2323f5e07da62bb7bed190