Set up thrift to hard fail if we get an http request with a head containing a newline
Summary:
The attached task has an immense amount of info. In short, in http/1.1 headers are newline terminated, so if you set a header containing a newline, you're actually setting two headers.

We have and probably still do set headers with user-controllable data as a way of data smuggling around the normal API. The task proves a method by which an attacker can do this to do nasty things.

Based on the two above points, the most sensical thing is just to ban setting headers with newlines in http/1.1

Prior to landing, we are going to be doing logging within SR to find any existing cases of this. I hope to discover there are exactly 0 (although I'll probably put one on purpose to prove logging works).

We add a KS in case we've somehow missed something.

Reviewed By: dmm-fb

Differential Revision: D46909419

fbshipit-source-id: 505c4dd6605717fb8052f7ae2512d15745adcaa0