Improve source file error handling in the compiler
Summary:
Make `source_manager::get_file` return an optional instead of throwing an exception when the file cannot be opened. This simplifies usage in the AST builder, avoids using exceptions for common(ish) control flow and gives the caller easy control of the reported diagnostic and location.

Also remove an unnecessary call to `get_file` in `parse_ast` and provide more details on how the API works in the comment of `get_file`.

Reviewed By: prasad223

Differential Revision: D47232205

fbshipit-source-id: 692a4dc1c821f2ec9794ac685872a317cf61f161