Move const value tests to compiler_test
Summary:
Move const value tests to the recently introduced compiler_test. The goal of these tests is to check that we produce diagnostics and therefore it is better be done as functional compiler tests rather than unit tests that neither check the exact diagnostics nor guarantee that the checker is run correctly.

The diagnostics themselves could be improved but keep them as is for now.

Reviewed By: Mizuchi

Differential Revision: D47300070

fbshipit-source-id: c91088856dcd94dc6d7c346f3d9da673964914f1