Split out terse intern box fixture to separate struct
Summary: we plan to optimize the empty check for intern box. The reason to change the fixture is that when the struct with unqualified and terse field together, the struct is never emptiable due to the unqualified field. This is just to change fixture so that we can have a fixture case where the struct is emptiable and uses intern box.

Reviewed By: vitaut

Differential Revision: D47414771

fbshipit-source-id: b02ed21a396a73dfa87861d826a8e7af388a1359