fix StructPatch::patchIfSet after clear
Summary:
What should the behavior of the following code be?

```
  MyStructPatch patch;
  patch.patchIfSet<ident::optI32Val>().clear();
  patch.patchIfSet<ident::optI32Val>() = 2;

  MyStruct obj;
  obj.optI32Val() = 1;
  patch.apply(obj);
```

It should be equivalent of the following code

```
  MyStructPatch patch1, patch2;
  patch1.patchIfSet<ident::optI32Val>().clear();
  patch2.patchIfSet<ident::optI32Val>() = 2;

  MyStruct obj;
  obj.optI32Val() = 1;
  patch1.apply(obj);
  patch2.apply(obj);
```

The `patch1` clear `optI32Val` field and the second patch is no-op (`patchIfSet` won't patch unset field). Thus the expected result would be

```
  assert(obj.optI32Val().has_value() == false);
```

However, without this diff we will get

```
  assert(obj.optI32Val() == 2);
```

Since

```
patch.patchIfSet<ident::optI32Val>() = 2;
```

will override

```
patch.patchIfSet<ident::optI32Val>().clear();
```

To fix the problem, if patch is already removed, we return the patch prior and clear the field in patch after (so that patch prior will be ignored).

Reviewed By: thedavekwon

Differential Revision: D47174290

fbshipit-source-id: 4c7630e19eb39dfa6c0f4ebdb118a853adf654e6