Move source range into definition attributes
Summary:
This is an intrinsic property of the definition and so is something we can include inline, even though it is only present in AST context (as e.g. runtime schema doesn't have access to the source code).
Also makes room for documentation source range which will be added in the next diff.

Reviewed By: vitaut

Differential Revision: D47560057

fbshipit-source-id: 0d00a1ef5d7f1bb808f61f169ea3a2b7e5a4249d