Revert D47530016: Multisect successfully blamed "D47530016: [Thrift] Use Intrinsic Default to construct and reset intern boxed fields does not have custom default. " for test or build failures
Summary:
This diff is reverting D47530016
D47530016: [Thrift] Use Intrinsic Default to construct and reset intern boxed fields does not have custom default.  by thedavekwon has been identified to be causing the following test or build failures:

Tests affected:
- [thrift/test/terse_write:terse_write_test - TerseWriteTest/TerseWriteSerializerTests/0.TerseInternBoxWithCustomDefault](https://www.internalfb.com/intern/test/562950033168781/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2573602
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: thedavekwon

Differential Revision: D47787087

fbshipit-source-id: deb2b2f71ae61e0797a30a5cf71eb604757cda77