Propagate enum constructors as part of a typedef
Summary:
This diff changes the Rust implementation of the Thrift statement `typedef somedependency.SomeEnum MyEnum`. Previously it would generate:

```
pub type MyEnum = somedependency::types::SomeEnum;
```

After this diff it instead generates:

```
pub use somedependency::types::SomeEnum as MyEnum;
```

Background reading: https://doc.rust-lang.org/1.71.0/reference/names/namespaces.html

In the **type namespace**, both implementations are equivalent. In the **value namespace**, the original defines nothing, whereas the new implementation exposes the constructor function for the underlying Rust tuple struct. Thrift enums in Rust are represented by:

```
// in somedependency

pub struct SomeEnum(pub i32);
```

which has a meaning in the type namespace (as expected; the struct's type) as well as the value namespace (a function with an argument of type `i32` that returns `SomeEnum`). The former is what makes `SomeEnum` legal to write in a function signature or a field of a struct definition. The latter is what makes `SomeEnum(-1)` legal to write in an expression.

For `typedef` in general, using `pub use` is not possible. `pub type` is correct. For example there is no `pub use` version of `put type MyArray = Vec<Thing>`.

For enums, since they cannot have a generic parameter in Thrift, `pub use` always works.

This ends up being necessary for {D47941012}.

Reviewed By: zertosh

Differential Revision: D47940993

fbshipit-source-id: 2832d2fd540abf4a1b721c5625a33deb1a288c1d