Make py3 reflection also give python field name
Summary: The thrift py converter relied on thrift py3 reflection, which is deprecated and wasn't really considered when we were doing the unstructured annotation migration, which broke the converter when using py3.name. Adding an extra field to py3 reflection which gives the py_name for the field to fix this.

Reviewed By: junzh0u

Differential Revision: D47933872

fbshipit-source-id: 3bf4f16333975921517c4d7f0a8520c8d0cb9f69