Improve handling of sinks and streams in the compiler
Summary:
Sink and stream are not types in the usual sense but grammatical constructs that associate various types and exceptions with functions. Move them out of the types vector for easier access and make them owned by containing functions. This also makes AST nodes closer to being correct by construction and moves some of the logic from the separate mutator phase to the AST building where it belongs. And finally this makes sinks and streams not being mixed with actual types in AST visitor which was a weird hack for visiting throws (!) clauses which obviously doesn't belong in templated type visitation.

Move `returns_sink` and `returns_stream` methods back to the public `t_function` API since they are pretty useful.

Reviewed By: iahs

Differential Revision: D47840589

fbshipit-source-id: a9972d531f9dbd8054a577d91cef8e600588c33c