Wrap FD received errors as TApplicationException
Summary:
All the other error paths in this file use `TApplicationException`, presumably because something up-stack catches either `TException` or `TApplicationException`.

When `popReceivedFdsFromSocket` was throwing `RocketException`, the net result was that the connection wasn't interrupted properly, and the client would time out.

With this change, the client immediately sees `(REQUEST_PARSING) Failed to parse request`, which is what we want.

Thanks to oandrew for pointing out the problem.

Differential Revision: D48014708

fbshipit-source-id: 6c03d76cda50b99426baf5c5c545b47b3d2bc18b