support serialized + marshal interop
Summary:
Some thrift files may be incompatible with marshal-based capi or may prefer serialization-based capi to reduce compile times. This diff makes that work.

When a marshal struct has a serialized struct field, it needs to store the internal_data representation of that struct. For this reason, the `ComposedStruct` `Constructor` uses the standalone `Constructor` to make a serialized object, then pulls out its `fbthrift_data` tuple and returns it for storage in the marshal struct's internal data tuple. This works because `getThriftData` implicitly increfs the tuple, ensuring it outlives the containing object which is discarded by `StrongRef`.

Reviewed By: Filip-F

Differential Revision: D48007603

fbshipit-source-id: 2cbc8ed93417ae407aeb01538a9019cbb0dfd524