add string_view ctor to mstch::node
Summary: It's safe to convert a valid `std::string_view` to `std::string`, but it involves a copy, so the conversion is generally explicit. This makes it slightly inconvenient / distracting to use `std::string_view` in mstch generators. To clean this up, add an implicit constructor of `mstch::node` from `std::string_view`.

Reviewed By: iahs

Differential Revision: D48123591

fbshipit-source-id: 390ac12babbc493780a5c695246ab72173eff9b7