thrift-rust: make optionals with defaults behavior in line with the thrift standard
Summary:
Thrift optional field with defaults should be initialized to `None` according to the standard. The current behavior is to initialize it to `Some`.

In order to transition to the new behavior (which is a breaking change), a thrift compiler flag `deprecated_optional_with_default_is_some` was introduced and compiling without it would trigger an error. In this stage, compilation error is removed and using optionals with defaults without the flag now generates spec-conforming code.

There are at least 3 scenarios where the default field value is being used, and they all have slightly different processing flows and behaviors for optional fields with defaults:
- `Default::default()`. In the old and the legacy behavior it initializes to `Some`. In the new one it is `None`. Flow: `structimpl.mustache` -> `fieldvalue.mustache` using `class rust_mstch_field`
- Constant values (and explicit defaults) when the field is omitted. In the old and the legacy behavior initializes to `Some`. In the new one it is `None`. Flow: `value.mustache` -> `fieldvalue.mustache` using `class mstch_rust_struct_field`
- `Deserialize::read`. Both old and new behavior is to use `None` if the value isn't present in the stream (or if set to `null`). Flow: `structimpl.mustache` -> `value.mustache` using `class rust_mstch_field`

Unfortunately, this change only further makes the three scenarios described above even more different. It might be desirable to follow up if these could be unified. It might be not trivial to do so though, especially while maintaining the not-so-consistent legacy behavior intact.

All the three flows are now covered by compatibility unit tests, both in current and legacy behavior.

There seem to be some pre-existing issues with thrift adapters compatibility with constants (and explicit defaults). This change shouldn't make it better nor worse. These compatibility issues should be addressed separately

Reviewed By: emersonford

Differential Revision: D46478604

fbshipit-source-id: d970cd6122db151303f305550536ab25ffa3fe70