Avoid inlining of std::string creation in Thrift Metadata
Summary:
There are a few cases where thrift's code generation has metadata related structures that take a std::string. This often inlines complex code in places that are only called once (because they are cached in Meyer's singletons).

This diff introduces overloads that take const char*s, reducing the repetitive generated code. Additionally, an uncommon destruction path of MethodMetadata is outlined so that it is not repeatedly inlined in generated code

Reviewed By: vitaut

Differential Revision: D48125905

fbshipit-source-id: 7fc76aa98be48a12a12cf035029fedc487319507