Use Semifuture to implement Future thrift methods
Summary: Under the hood, the current implementation of futures appears to be the same as getting a SemiFuture and calling `toUnsafeFuture`. Would appreciate experts on the future API confirming this is indeed the correct thing to do. There's a bunch of dead code that can be removed after this which I'll save for a followup to reduce the risk of having to roll back.

Reviewed By: yfeldblum

Differential Revision: D48194024

fbshipit-source-id: 9916091a80dd16665a10cd3c4d99f6e043454bc1