writeChainWithFds no longer FATALs when FDs are in `Received` state
Summary:
Currently, Thrift FD requires any handler that receives FDs to `fds->releaseReceived()` before returning. Failing to do so currently fatals in`AsyncFdSocket::writeChainWithFds` when it tries to `releaseToSend`.

This can be bad in a production service. Specifically, an errant client that sends FDs to a method not expecting them can crash the server. Instead, it is better that the client should get an error, and the server cloe the connection.

This issue only concerns the server's send code path, and only for request-response methods. For both client sends and stream sends, the presence of "received" instead of "to send" FDs at the linked callsite **could** remain a hard failure.

This changes `AsyncFdSocket::writeChainWithFds` to `failWrite(BAD_ARGS)` upon seeing `fds` containing "received FDs" instead of "FDs to send". This will make this error non-fatal for **all** the send code-paths, not just server response. On the plus side, `failWrite` will fail the connection, which seems desirable.

Alternatives considered and rejected:
  - Update `packWithFds` to DFATAL and put the payload into an error state (which will cause `writeChainWithFds` to `failWrite` as above). The upside here is that, if desired, this could distinguish "server response" from the other code paths, and `DFATAL` unless on the server path. The downside is that distinguishing the code paths, while done via templates, would still require careful perf testing, which I'm not keen to repeat.
 - REJECTED: `writeChainWithFds` DFATALs and converts `SocketFds::Received` to `SocketFds::ToSend`. I don't like that this turns an error condition into a success, and wastes FD resources to boot.
- REJECTED: `ThriftServerRequestResponse::sendThriftResponse` tests for the error condition. This is DEFINITELY adding branches and new code on the hot path, so it's my least favorite approach.
 - Your idea here?

For all discussed solutions, in the scenario that a client attaches FDs to a request that doesn't expect them, the server method handle will run, and unnecessarily consume server resources for processing -- before erroring out. However, doing better would require annotating which methods can receive FDs, and I don't want to open this can of worms.

Reviewed By: robertroeser

Differential Revision: D48287769

fbshipit-source-id: d2b7e9446b889f5e34af20b81392926b74056f9e