reserve `size_of(size_t)` bytes at the front of buffer to encode buffer size in AllocatingParserStrategy
Summary:
Previously, we need to pass in
```
auto* allocAndSize =
        new std::pair<Allocator&, size_t>(allocator_, currentBufferSize_);
```
to `deallocate` method. This introduced un-necessary heap allocation. In order to get rid of this additional allocation, we can use the first `size_of(size_t)` bytes of the buffer to encode the length of the buffer. This is possible since once allocated, buffer length is never going to change until deallocated.

We also shift all pointer arithmetic by `size_of(size_t)` to compensate the fact that first `size_of(size_t)` bytes are already used for bookkeeping.

Reviewed By: robertroeser

Differential Revision: D48325342

fbshipit-source-id: 517bf061d6505611dc375b0721762490e631fae8