Implement Thrift Flags support for strings
Summary:
This diff adds supports for `std::string` for Thrift Flags. Most of the changes are just plumbing for a newly added type (template specialization, overloads, new data members etc.)

The majority of the complexity added comes from changes to how the flag observers store data. To optimize read performance, `bool` and `int64_t` use `ReadMostlyAtomicObserver`. This does not work for `std::string` since it's not trivially copyable (cannot be stored in `std::atomic`).

Other than that, the behavior is equivalent between `std::string` and existing types. This means that reading the `std::string` implies a copy. This is a trade-off to keep things simple and consistent with the existing trivially copyable types. The assumption here is that `std::string`-based flags won't be read directly via `get()` (because string comparisons are expensive anyways) and instead will be accessed via `observe` and preprocessed into more read-friendly data structures.

Reviewed By: avalonalex, sazonovkirill

Differential Revision: D48375420

fbshipit-source-id: edc254f299de73a5af819ed73759e557fbc35021