(rocket framing) fix potential double free in AllocatingParserStrategy
Summary:
in `AllocatingParserStrategy`, if `owner_.handleFrame` throws exception (for example, when frame is bogus), `buffer_` object will be deallocated twice. Once when destructor of `IOBuf` is called and once when destructor of `AllocatingParserStrategy` is called.

In order to fix this, we need to introduce a staging buffer to be hand over to the owner and not let `AllocatingParserStrategy` go into inconsistent state even if `owner_.handleFrame` throws exception

Reviewed By: robertroeser

Differential Revision: D48451742

fbshipit-source-id: 7f14c86e9c509d7d20c92db5d38e6104204495e7