generate structs for adapted underlying types
Summary: The current implementation generates structs based on the struct's cpp name, but for adapted structs, the cpp name might not actually be a thrift type. To make it easier to support field-adapted types in next diff, generate the struct for the actual underlying thrift type cpp2 struct type, which can then be converted into the adapted type as needed.

Reviewed By: vitaut

Differential Revision: D48009698

fbshipit-source-id: 6548b7b60aab744e7d68c3d975b006cd434d4f5a