Fix the qualified identifier doc
Summary: The documentation of a qualified identifier gives an impression that the first component is always a file name while in fact it can be a definition name in case of enums. Fix the wording to not exclude the enum case. The documentation for enums still needs some cleanup which will be done separately.

Reviewed By: thedavekwon

Differential Revision: D48659018

fbshipit-source-id: e786f796f9800fcf858246b3e20d52bc6ca07750