split up NaN conformance tests
Summary:
Move the Conformance tests with NaN to a separate section, since they aren't a Go specific issue.
This will allow us to concentrate our efforts on Go specific issues.

I also Moved list<float> and list<double> tests to NaN section

```
$ buck2 test 'fbcode//mode/dev' fbcode//thrift/conformance/test/data:round-trip-test -- --exact 'thrift/conformance/test/data:round-trip-test - go/RoundTripTest/testset.list<float>/all.Binary'
```
I saw the Actual and Expected are the same and it includes a NaN
I assume that all list of float tests include a NaN so also moved them to NaN section as they are also in CPP skip section

Reviewed By: thedavekwon

Differential Revision: D48606306

fbshipit-source-id: 6ef25ca495a0d55aff04dab7388c35428bb4a2bc