Optimize apache::thrift::protocol::detail::ApplyPatch::operator()
Summary:
**Created From Opportunity Instance: **[**211729598278073**](/overwatchx/opportunities/node?opportunity_instance_id=211729598278073)

**Background/help** for std::unordered_set: [https://www.internalfb.com/intern/servicedoctor/help/#std%3A%3Aunordered_set%2CCPU](/intern/servicedoctor/help/#std%3A%3Aunordered_set%2CCPU)

**BigGrep**: https://fburl.com/code/pafet8qj

**Strobelight Icicle**: [https://fburl.com/scuba/strobelight_services/fx73f6y9](https://fburl.com/scuba/strobelight_services/fx73f6y9)

**Strobelight GraphProfiler**: [https://fburl.com/scuba/strobelight_services/5b3p45xy](https://fburl.com/scuba/strobelight_services/5b3p45xy)

**Strobelight Leaf File:Line**: [https://fburl.com/scuba/strobelight_services/wpk146m8](https://fburl.com/scuba/strobelight_services/wpk146m8)

**Call Graph**: https://fburl.com/cv/0xupdmb9

**Strobelight CPU**: [https://fburl.com/scuba/strobelight_services/g6zvurw9](https://fburl.com/scuba/strobelight_services/g6zvurw9)

**Rev f5d852cf46**: [Patch.cpp:400,329,239](https://fburl.com/code/ko9x27jp)

**Rev bd85c06ef5**: [Patch.cpp:400,329,239,397](https://fburl.com/code/zffb9x8j)

**Rev 5e6930a86b**: [Patch.cpp:400,329,239,397](https://fburl.com/code/t7r04gqv)

As of 2023-08-24, the expected CPU win for optimizing this function is **0%**

_This Task was filed via _[_ServiceDoctor_](/intern/servicedoctor/)_ to help minimize work duplication via advertising perf investigations. As such, please keep the string __`"apache::thrift::protocol::detail::ApplyPatch::operator()"`__ in the contents of this Task so ServiceDoctor knows about it (even if this Task is closed)._

=====================[**INSTRUCTIONS FOR STATE CHANGE**]=====================

**Everything Has Done? Record Win To Mark It As Completed!**
You can navigate back to the [opportunity instance page](/overwatchx/opportunities/node?opportunity_instance_id=211729598278073) and record the win through the button in the action section.
Or you can directly close the task and the bot will reply on the task about how to record a win.
After you recorded the win, the task will be auto-closed and opportunity will be marked as COMPLETED.

**Invalid Opportunity? Blocklist It With Feedback!**
If you think this opportunity is not a valid opportunity, you could try to blocklist it through the action section in [opportunity instance page](/overwatchx/opportunities/node?opportunity_instance_id=211729598278073).
You will be asked to provide comments, blocklist category and a suspend time according to your assessment.
Your feedback will help us keep this community in a good shape.

**No Longer Available? Return This Opportunity To The Pool!**
If you think you no longer want to work on this task, you can change the owner to up to grab.
This will automatically return this opportunity to the pool for other people to take it over.

Notes: Trying to explore a mechanism to use bitsets for the fact that we only really have 10 possible enum values to pass so we can leverage a bitset here - TBH, a 10 bit representation would suffice for this and make this a super-lightweight operation

Differential Revision: D48684797

fbshipit-source-id: cc568afaed0824495a239da9d421bba14e06b15f