Add createdInteraction to ServiceRequestInfoMap
Summary: We already have `interactionName` in `ServiceRequestInfoMap`. If a regular RPC creates an interaction we need that info in MethodMetadata so that we can select the right generated method to call for an interaction. In this diff we add an extra field in `ServiceRequestInfoMap` in the generated code that will be the name of the created interaction if that RPC creates an interaction and `std::nullopt` otherwise.

Reviewed By: sazonovkirill

Differential Revision: D47907585

fbshipit-source-id: d06301532f336a268494cfbef4e0d8becce5e196